requirements.txt
# Core web framework
Flask==3.1.0
flask-cors==5.0.1
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0

# Computer Vision and AI
opencv-python==4.11.0.86
opencv-contrib-python==4.11.0.86
ultralytics==8.3.86
torch==2.6.0
torchvision==0.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
dlib==19.24.6

# Data Processing
numpy==2.1.1
pandas==2.2.3
pillow==10.2.0
matplotlib==3.10.1
seaborn==0.13.2

# Database
SQLAlchemy==2.0.39
psycopg2-binary==2.9.10

# Utilities
python-dotenv==1.0.1
requests==2.31.0
imutils==0.5.4
tqdm==4.67.1
plyer==2.1.0
tabulate==0.9.0
watchdog==6.0.0

# Development and Testing
black==25.1.0
pytest
pytest-cov

# Optional: For advanced features
# streamlit==1.43.2  # Uncomment if using Streamlit interface
# tensorboard==2.19.0  # Uncomment if using TensorBoard for training monitoring
